Allow malloc use 1% of its power
Up to 4gig per call (or max heap, whatever happens first :)

This will allow Userland map the whole screen in a single memory location (~1mb if 1280x768)

+++ Kernel/memory/memory.h
@@ -3,8 +3,8 @@
 
 #include <memory.h>
 
-typedef uint16_t heap_entry;
-typedef uint16_t map_entry;
+typedef uint32_t heap_entry;
+typedef uint32_t map_entry;
 
 // 256 MB
 #define MAP_START 0x10000000
